websocket_backend/websocket_project/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
eventlet==0.38.2
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pymongo==4.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
redis==5.2.1
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
